refactor(oidc): split org_keys into a directory module
Convert services/oidc/org_keys.rs (2,037 lines) into a directory module
by concern:

- mod.rs: module docs, org_issuer_or_base, the OrgKeys/OrgKeySetSnapshot
  types, and the key-material substrate (KeyMaterial, generate, ensure,
  build_snapshot, state_priority) shared by the children, plus the
  test_support fixtures (setup, backdate, NO_OPERATOR)
- resolve.rs: the hot read path — OrgKeysCache (1-min TTL) and
  resolve_org_keys, org_jwks
- rotation.rs: operator rotate/revoke and the emergency path together
  with their timing gates and outcome types (they share the error type,
  constants, and audit helper); the gate helpers are pub(super) for the
  panel
- panel.rs: the admin-page signing-key view-model

The 21 inline tests move with the code they test. Pure moves — no
renamed types, no behavior change; services::oidc:: paths are unchanged
so all consumers (AppState, discovery, admin handlers, vouch-tests) are
untouched. The rotate and emergency transaction helpers stay separate
per the issue scope.

Closes #641
