+++ crates/vouch-server/src/services/oidc/org_keys/mod.rs
@@ -1,0 +1,496 @@
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+//! Per-organization OIDC issuer signing keys and rotation.
+//!
+//! Each org that has claimed an issuer subdomain signs its OIDC federation
+//! tokens with its **own** keys, served only at its own JWKS — so the issuer
+//! host is a real cryptographic tenant boundary (a token for org A does not
+//! verify against org B's keys). This is generic OIDC issuer isolation, the
+//! same property Okta/Auth0/Ory give each tenant; it lets any relying party
+//! (AWS STS/Identity Center, GCP/Azure workload identity, Kubernetes, Vault,
+//! or a plain OIDC RP) trust the org's issuer without a per-token condition.
+//!
+//! An org's keys are a **set** ([`OrgKeys`]): one ES256 and one RS256, created
+//! and loaded together — a per-org issuer always has both. [`resolve_org_keys`]
+//! is the single entry point for signing; callers pick the algorithm they need
+//! and fall back to the common platform key when the org has none. The key
+//! material is generated by aws-lc-rs (via [`OidcSigningKey`] /
+//! [`OidcRsaSigningKey`]) and sealed at rest by the document store.
+//!
+//! ## Rotation lifecycle (Auth0 model)
+//!
+//! A key set always holds an [`SigningKeyState::Current`] signer **and** a
+//! pre-staged [`SigningKeyState::Next`] successor: both are created on first
+//! use and the successor is re-staged in the same transaction that consumes
+//! it, so the org JWKS permanently publishes the key that will sign next and
+//! relying-party caches (AWS IAM caches JWKS on an undocumented schedule) are
+//! always pre-warmed. Nothing transitions on a timer — both rotation steps are
+//! explicit operator actions:
+//!
+//! 1. **Rotate** (`rotate_org_keys`, `POST /admin/subdomain/rotate`): promotes
+//!    Next → Current, demotes the old signer to [`SigningKeyState::Previous`]
+//!    (published, verify-only), and stages a fresh Next — both algorithms in
+//!    one transaction. Rejected until the Next keys have been published for
+//!    [`rotation::PUBLISH_AHEAD_HOURS`] (cache warm-up) and while an unrevoked Previous
+//!    exists.
+//! 2. **Revoke** (`revoke_org_previous_keys`, `POST /admin/subdomain/revoke`):
+//!    deletes the Previous keys. Rejected until the token-drain window
+//!    (`max(session_hours, floor) + margin`) since the demoting rotate has
+//!    elapsed, so no outstanding token loses its verification key.
+//!
+//! The emergency path (`emergency_rotate_org_keys`) replaces the entire key
+//! set at once: new Current and Next for both algorithms, Previous deleted
+//! outright, in one transaction. Outstanding tokens signed by the compromised
+//! keys will fail verification until RPs refetch the JWKS.
+//!
+//! ## Cache invariant
+//!
+//! [`OrgKeysCache`] holds a **unified snapshot** per org: both the active
+//! signers (for token issuance) and the ordered public-JWK list (for the JWKS
+//! endpoint) are projections of the same `list_all("org_id")` result, so the
+//! signing key is always in this instance's own published JWKS. Every rotation
+//! transition calls `invalidate(org_id)` so the next request rebuilds from the
+//! DB immediately.
+//!
+//! ## Encryption guard
+//!
+//! Keys are created on first use, but **only when the document store encrypts
+//! at rest** ([`crate::db::store::DocumentStore::is_encrypted`]). In the dev
+//! plaintext store no key is created and the resolver returns `None`, so
+//! callers fall back to the common key — a private key is never persisted in
+//! the clear, and there is no config flag threaded through call sites.
+
+use anyhow::{Context, Result};
+use axum::http::StatusCode;
+use base64::Engine;
+use base64::engine::general_purpose::STANDARD;
+use jiff::Timestamp;
+use secrecy::ExposeSecret;
+use zeroize::Zeroizing;
+
+use crate::config::ServerConfig;
+use crate::db::document_type::Document;
+use crate::db::documents::oauth::JwsAlgorithm;
+use crate::db::documents::organization::{OrgSigningKeyDoc, SigningKeyState};
+use crate::db::store::DocumentStore;
+use crate::db::{self, Organization};
+use crate::error::ServiceError;
+
+use crate::crypto::keys::{Jwk, OidcRsaSigningKey, OidcSigningKey};
+
+mod panel;
+mod resolve;
+mod rotation;
+
+pub(crate) use panel::{OrgKeyPanel, org_key_panel};
+pub use resolve::{OrgKeysCache, org_jwks, resolve_org_keys};
+pub use rotation::{
+    EmergencyOutcome, Operator, RevokeOutcome, RotateOutcome, emergency_rotate_org_keys,
+    revoke_org_previous_keys, rotate_org_keys,
+};
+
+/// The issuer an org's federation tokens must use: its claimed subdomain
+/// issuer, or the shared base URL when it has no subdomain.
+///
+/// **Fail-closed:** an org that claimed a subdomain must never fall back to the
+/// shared issuer — its tokens would not match its discovery document or the
+/// relying party's OIDC provider config. A subdomain that can't be turned into
+/// an issuer URL is an error, not a silent downgrade.
+///
+/// Issuer selection is gated on whether a subdomain is *claimed*, not on
+/// at-rest encryption — unlike [`resolve_org_keys`], which yields a per-org
+/// signing key only under encryption. These stay consistent because claiming
+/// requires an encrypted store (enforced at the claim handler and the startup
+/// guard) and `is_encrypted()` is fixed for the life of the process: a claimed
+/// subdomain therefore implies per-org keys exist, so the org issuer is always
+/// backed by an org-specific key. The one exception — org issuer served under
+/// the shared key — requires an out-of-band DB write on a plaintext deployment;
+/// [`resolve_org_keys`]'s encryption check is the fail-safe backstop for it.
+///
+/// # Errors
+/// Returns `ServiceError` if a claimed label can't be built into an issuer URL.
+pub fn org_issuer_or_base(
+    config: &ServerConfig,
+    org: Option<&Organization>,
+) -> Result<String, ServiceError> {
+    if let Some(label) = org.and_then(|o| o.subdomain.as_deref()) {
+        config.org_issuer(label).ok_or_else(|| {
+            tracing::error!("could not build org issuer for label '{label}' from base_url");
+            ServiceError::api(
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "issuer_error",
+                "Failed to construct the organization issuer",
+            )
+        })
+    } else {
+        Ok(config.base_url.clone())
+    }
+}
+
+/// An org's active signing keys — one per algorithm.
+///
+/// Callers project the field they need (`&keys.es256` / `&keys.rs256`).
+pub struct OrgKeys {
+    pub es256: OidcSigningKey,
+    pub rs256: OidcRsaSigningKey,
+}
+
+/// A unified snapshot of an org's full key set, held by [`OrgKeysCache`].
+///
+/// Both projections (signing and JWKS) derive from the same `list_all("org_id")`
+/// result, so the signing key is always in this instance's own published JWKS —
+/// no cross-projection staleness within a single server instance.
+pub struct OrgKeySetSnapshot {
+    /// Current signers for token issuance — ES256 and RS256.
+    pub signers: OrgKeys,
+    /// Ordered public JWKs for all live keys (Current + Next + Previous).
+    ///
+    /// Sorted: all RS256 keys first (Current → Next → Previous),
+    /// then all ES256 keys in the same order. Per OIDC Core §3.1.3.7 RS256
+    /// must appear first; within an algorithm the order is deterministic for
+    /// stable response bytes and test assertions.
+    pub jwks: Vec<Jwk>,
+}
+
+/// Freshly generated key material for one algorithm: PKCS#8 DER + derived kid.
+struct KeyMaterial {
+    der: Zeroizing<Vec<u8>>,
+    kid: String,
+}
+
+impl KeyMaterial {
+    /// Build an [`OrgSigningKeyDoc`] carrying this material in `state`.
+    ///
+    /// Borrows so the same material can be reused across OCC retries.
+    fn doc(&self, org_id: &str, alg: JwsAlgorithm, state: SigningKeyState) -> OrgSigningKeyDoc {
+        OrgSigningKeyDoc {
+            org_id: org_id.to_string(),
+            alg,
+            kid: self.kid.clone(),
+            private_pkcs8_der_b64: STANDARD.encode(&self.der).into(),
+            state,
+            staged_at: None,
+            demoted_at: None,
+        }
+    }
+}
+
+/// Generate fresh key material for `alg`.
+///
+/// RSA generation is expensive, so it runs on the blocking pool and callers
+/// keep it out of retry loops.
+async fn generate_key_material(alg: JwsAlgorithm) -> Result<KeyMaterial> {
+    match alg {
+        JwsAlgorithm::Es256 => {
+            let der = OidcSigningKey::generate_pkcs8_der().context("ES256 keygen")?;
+            let kid = OidcSigningKey::from_pkcs8_der(&der)
+                .context("ES256: parse DER")?
+                .key_id()
+                .to_string();
+            Ok(KeyMaterial { der, kid })
+        }
+        JwsAlgorithm::Rs256 => {
+            let der = tokio::task::spawn_blocking(OidcRsaSigningKey::generate_pkcs8_der)
+                .await
+                .context("RS256 keygen task panicked")??;
+            let kid = OidcRsaSigningKey::from_pkcs8_der(&der)
+                .context("RS256: parse DER")?
+                .key_id()
+                .to_string();
+            Ok(KeyMaterial { der, kid })
+        }
+        other => anyhow::bail!("generate_key_material: unsupported algorithm {other:?}"),
+    }
+}
+
+/// Create the key in `state` for `(org_id, alg)` if it doesn't exist.
+///
+/// `Next` keys record `staged_at = now` — the start of their publish window
+/// (the always-staged invariant: the successor is published in the JWKS from
+/// first use so relying-party caches are warm long before any rotate).
+/// Idempotent: the deterministic ID makes a concurrent creation or retry
+/// collide on the primary key rather than insert a duplicate.
+async fn ensure_key(
+    store: &DocumentStore,
+    org_id: &str,
+    alg: JwsAlgorithm,
+    state: SigningKeyState,
+) -> Result<()> {
+    if state == SigningKeyState::Previous {
+        anyhow::bail!("ensure_key: Previous keys are only created by rotation");
+    }
+    if db::get_org_signing_key(store, org_id, alg, state)
+        .await?
+        .is_some()
+    {
+        return Ok(());
+    }
+    let doc = OrgSigningKeyDoc {
+        staged_at: (state == SigningKeyState::Next).then(Timestamp::now),
+        ..generate_key_material(alg).await?.doc(org_id, alg, state)
+    };
+    db::try_insert_org_signing_key(store, &doc).await?;
+    Ok(())
+}
+
+/// Decode the DER from an `OrgSigningKeyDoc` into a `Zeroizing<Vec<u8>>`.
+fn decode_der(doc: &OrgSigningKeyDoc) -> Result<Zeroizing<Vec<u8>>> {
+    Ok(Zeroizing::new(
+        STANDARD
+            .decode(doc.private_pkcs8_der_b64.expose_secret())
+            .context("decode org signing key DER")?,
+    ))
+}
+
+/// State-priority ordering used for JWKS sort within an algorithm.
+fn state_priority(state: SigningKeyState) -> u8 {
+    match state {
+        SigningKeyState::Current => 0,
+        SigningKeyState::Next => 1,
+        SigningKeyState::Previous => 2,
+    }
+}
+
+/// Build a unified [`OrgKeySetSnapshot`] from a pre-loaded slice of key docs.
+///
+/// Finds the Current ES256 and RS256 keys for signing, and builds the sorted
+/// public-JWK list for the JWKS endpoint. Returns `None` if no Current key is
+/// found for either algorithm.
+///
+/// **Ordering:** RS256 keys first, then ES256; within each algorithm:
+/// Current → Next → Previous. This is a hard constraint for OIDC Core
+/// §3.1.3.7 compliance (RS256 must appear first).
+///
+/// Takes pre-loaded docs from the caller (avoids an extra DB round-trip
+/// when `resolve_org_keys` already has the full doc list in hand).
+fn build_snapshot(docs: &[Document<OrgSigningKeyDoc>]) -> Result<Option<OrgKeySetSnapshot>> {
+    // Separate by alg and sort by state priority within each alg.
+    let mut rs256_docs: Vec<_> = docs
+        .iter()
+        .filter(|d| d.data.alg == JwsAlgorithm::Rs256)
+        .collect();
+    let mut es256_docs: Vec<_> = docs
+        .iter()
+        .filter(|d| d.data.alg == JwsAlgorithm::Es256)
+        .collect();
+    rs256_docs.sort_by_key(|d| state_priority(d.data.state));
+    es256_docs.sort_by_key(|d| state_priority(d.data.state));
+
+    // Find the Current key for each alg — none means the org has no live
+    // signing capability; return None to fall back to the common key.
+    let Some(current_rs256_doc) = rs256_docs
+        .iter()
+        .find(|d| d.data.state == SigningKeyState::Current)
+    else {
+        return Ok(None);
+    };
+    let Some(current_es256_doc) = es256_docs
+        .iter()
+        .find(|d| d.data.state == SigningKeyState::Current)
+    else {
+        return Ok(None);
+    };
+
+    let rs256_der = decode_der(&current_rs256_doc.data)?;
+    let es256_der = decode_der(&current_es256_doc.data)?;
+    let rs256 = OidcRsaSigningKey::from_pkcs8_der(&rs256_der)?;
+    let es256 = OidcSigningKey::from_pkcs8_der(&es256_der)?;
+
+    // Build public JWKs: RS256 first (all states), then ES256 (all states).
+    let mut jwks = Vec::with_capacity(rs256_docs.len().saturating_add(es256_docs.len()));
+    for doc in &rs256_docs {
+        let der = decode_der(&doc.data)?;
+        let key = OidcRsaSigningKey::from_pkcs8_der(&der)?;
+        jwks.push(Jwk::Rsa(key.public_key_jwk()?));
+    }
+    for doc in &es256_docs {
+        let der = decode_der(&doc.data)?;
+        let key = OidcSigningKey::from_pkcs8_der(&der)?;
+        jwks.push(Jwk::Ec(key.public_key_jwk()?));
+    }
+
+    Ok(Some(OrgKeySetSnapshot {
+        signers: OrgKeys { es256, rs256 },
+        jwks,
+    }))
+}
+
+#[cfg(test)]
+#[expect(
+    clippy::unwrap_used,
+    clippy::expect_used,
+    clippy::panic,
+    reason = "test code: panic on assertion failure is acceptable"
+)]
+pub(super) mod test_support {
+    use std::sync::Arc;
+
+    use jiff::{Span, Timestamp};
+
+    use super::Operator;
+    use crate::db::documents::organization::SigningKeyState;
+    use crate::db::{
+        JwsAlgorithm, OrgSigningKeyDoc, Organization, claim_subdomain, deterministic_org_key_id,
+    };
+    use crate::test_utils::{create_test_org, test_app_state_encrypted};
+
+    /// No operator identity — service-level tests don't exercise audit trails.
+    pub(super) const NO_OPERATOR: Operator<'static> = Operator {
+        user_id: None,
+        email: None,
+    };
+
+    /// Encrypted AppState with a fresh org holding a claimed subdomain.
+    pub(super) async fn setup() -> (Arc<crate::AppState>, String, Organization) {
+        let state = test_app_state_encrypted().await;
+        let org = create_test_org(&state.store, "acme.com").await;
+        claim_subdomain(&state.store, &org.id, "acme-com")
+            .await
+            .unwrap();
+        let org = crate::db::get_organization(&state.store, &org.id)
+            .await
+            .unwrap()
+            .unwrap();
+        let org_id = org.id.clone();
+        (state, org_id, org)
+    }
+
+    /// Rewrite a key's publish/demotion timestamp so gate tests don't sleep.
+    pub(super) async fn backdate(
+        state: &crate::AppState,
+        org_id: &str,
+        alg: JwsAlgorithm,
+        key_state: SigningKeyState,
+        hours: i64,
+    ) {
+        let id = deterministic_org_key_id(org_id, alg, key_state);
+        let doc = state
+            .store
+            .get::<OrgSigningKeyDoc>(&id)
+            .await
+            .unwrap()
+            .expect("key to backdate must exist");
+        let past = Timestamp::now()
+            .checked_sub(Span::new().hours(hours))
+            .unwrap();
+        let mut data = doc.data;
+        match key_state {
+            SigningKeyState::Next => data.staged_at = Some(past),
+            SigningKeyState::Previous => data.demoted_at = Some(past),
+            SigningKeyState::Current => panic!("current keys have no timestamp"),
+        }
+        state.store.update(&id, &data).await.unwrap();
+    }
+}
+
+#[cfg(test)]
+#[expect(
+    clippy::unwrap_used,
+    clippy::expect_used,
+    reason = "test code: panic on assertion failure is acceptable"
+)]
+mod tests {
+    use jiff::Timestamp;
+
+    use super::{org_issuer_or_base, state_priority};
+    use crate::db::documents::organization::SigningKeyState;
+    use crate::db::{JwsAlgorithm, Organization, deterministic_org_key_id};
+    use crate::test_utils::test_config;
+
+    fn org(subdomain: Option<&str>) -> Organization {
+        Organization {
+            id: "org1".to_string(),
+            domain: "acme.com".to_string(),
+            name: None,
+            created_at: Timestamp::now(),
+            created_by_user_id: None,
+            additional_domains: Vec::new(),
+            subdomain: subdomain.map(String::from),
+        }
+    }
+
+    #[test]
+    fn issuer_is_org_subdomain_or_base_url() {
+        let config = test_config();
+        // No org, or an org without a subdomain -> shared base URL.
+        assert_eq!(org_issuer_or_base(&config, None).unwrap(), config.base_url);
+        assert_eq!(
+            org_issuer_or_base(&config, Some(&org(None))).unwrap(),
+            config.base_url
+        );
+        // A claimed subdomain -> the org issuer host.
+        assert_eq!(
+            org_issuer_or_base(&config, Some(&org(Some("acme")))).unwrap(),
+            "https://acme.test.example.com"
+        );
+    }
+
+    #[test]
+    fn state_priority_orders_current_next_previous() {
+        assert!(state_priority(SigningKeyState::Current) < state_priority(SigningKeyState::Next));
+        assert!(state_priority(SigningKeyState::Next) < state_priority(SigningKeyState::Previous));
+    }
+
+    #[test]
+    fn org_key_state_serde_round_trip_and_legacy_default() {
+        for state in [
+            SigningKeyState::Current,
+            SigningKeyState::Next,
+            SigningKeyState::Previous,
+        ] {
+            let json = serde_json::to_string(&state).expect("serialise");
+            let back: SigningKeyState = serde_json::from_str(&json).expect("deserialise");
+            assert_eq!(state, back, "round-trip failed for {json}");
+        }
+        // Rows written before rotation existed have no `state` field and must
+        // deserialize as Current.
+        assert_eq!(SigningKeyState::default(), SigningKeyState::Current);
+    }
+
+    /// The Current ID must be byte-identical to the deterministic ID used
+    /// before rotation existed so old rows keep their IDs without migration.
+    #[test]
+    fn current_id_matches_pre_rotation_formula() {
+        use aws_lc_rs::digest::{self, SHA256};
+
+        let org_id = "migration-parity-test-org";
+        let alg = JwsAlgorithm::Es256;
+
+        let mut ctx = digest::Context::new(&SHA256);
+        ctx.update(b"org_signing_key\0");
+        ctx.update(org_id.as_bytes());
+        ctx.update(b"\0");
+        ctx.update(alg.as_str().as_bytes());
+        let expected = hex::encode(ctx.finish().as_ref());
+
+        assert_eq!(
+            deterministic_org_key_id(org_id, alg, SigningKeyState::Current),
+            expected
+        );
+    }
+
+    #[test]
+    fn key_ids_are_distinct_per_state_alg_and_org() {
+        for alg in [JwsAlgorithm::Es256, JwsAlgorithm::Rs256] {
+            let ids = [
+                deterministic_org_key_id("test-org", alg, SigningKeyState::Current),
+                deterministic_org_key_id("test-org", alg, SigningKeyState::Next),
+                deterministic_org_key_id("test-org", alg, SigningKeyState::Previous),
+            ];
+            let unique: std::collections::HashSet<&String> = ids.iter().collect();
+            assert_eq!(unique.len(), 3, "{alg:?}: state IDs collide: {ids:?}");
+            // Deterministic: the same inputs produce the same ID.
+            assert_eq!(
+                deterministic_org_key_id("test-org", alg, SigningKeyState::Current),
+                ids[0]
+            );
+        }
+        assert_ne!(
+            deterministic_org_key_id("org-a", JwsAlgorithm::Es256, SigningKeyState::Current),
+            deterministic_org_key_id("org-b", JwsAlgorithm::Es256, SigningKeyState::Current),
+        );
+    }
+
+    // -----------------------------------------------------------------------
+    // Async tests against an encrypted store
+    // -----------------------------------------------------------------------
+}
