+++ crates/vouch-server/src/services/oidc/org_keys/rotation.rs
@@ -1,88 +1,21 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
-//! Per-organization OIDC issuer signing keys and rotation.
+//! Operator-driven rotation of per-org issuer signing keys.
 //!
-//! Each org that has claimed an issuer subdomain signs its OIDC federation
-//! tokens with its **own** keys, served only at its own JWKS — so the issuer
-//! host is a real cryptographic tenant boundary (a token for org A does not
-//! verify against org B's keys). This is generic OIDC issuer isolation, the
-//! same property Okta/Auth0/Ory give each tenant; it lets any relying party
-//! (AWS STS/Identity Center, GCP/Azure workload identity, Kubernetes, Vault,
-//! or a plain OIDC RP) trust the org's issuer without a per-token condition.
-//!
-//! An org's keys are a **set** ([`OrgKeys`]): one ES256 and one RS256, created
-//! and loaded together — a per-org issuer always has both. [`resolve_org_keys`]
-//! is the single entry point for signing; callers pick the algorithm they need
-//! and fall back to the common platform key when the org has none. The key
-//! material is generated by aws-lc-rs (via [`OidcSigningKey`] /
-//! [`OidcRsaSigningKey`]) and sealed at rest by the document store.
-//!
-//! ## Rotation lifecycle (Auth0 model)
-//!
-//! A key set always holds an [`SigningKeyState::Current`] signer **and** a
-//! pre-staged [`SigningKeyState::Next`] successor: both are created on first
-//! use and the successor is re-staged in the same transaction that consumes
-//! it, so the org JWKS permanently publishes the key that will sign next and
-//! relying-party caches (AWS IAM caches JWKS on an undocumented schedule) are
-//! always pre-warmed. Nothing transitions on a timer — both rotation steps are
-//! explicit operator actions:
-//!
-//! 1. **Rotate** (`rotate_org_keys`, `POST /admin/subdomain/rotate`): promotes
-//!    Next → Current, demotes the old signer to [`SigningKeyState::Previous`]
-//!    (published, verify-only), and stages a fresh Next — both algorithms in
-//!    one transaction. Rejected until the Next keys have been published for
-//!    [`PUBLISH_AHEAD_HOURS`] (cache warm-up) and while an unrevoked Previous
-//!    exists.
-//! 2. **Revoke** (`revoke_org_previous_keys`, `POST /admin/subdomain/revoke`):
-//!    deletes the Previous keys. Rejected until the token-drain window
-//!    (`max(session_hours, floor) + margin`) since the demoting rotate has
-//!    elapsed, so no outstanding token loses its verification key.
-//!
-//! The emergency path (`emergency_rotate_org_keys`) replaces the entire key
-//! set at once: new Current and Next for both algorithms, Previous deleted
-//! outright, in one transaction. Outstanding tokens signed by the compromised
-//! keys will fail verification until RPs refetch the JWKS.
-//!
-//! ## Cache invariant
-//!
-//! [`OrgKeysCache`] holds a **unified snapshot** per org: both the active
-//! signers (for token issuance) and the ordered public-JWK list (for the JWKS
-//! endpoint) are projections of the same `list_all("org_id")` result, so the
-//! signing key is always in this instance's own published JWKS. Every rotation
-//! transition calls `invalidate(org_id)` so the next request rebuilds from the
-//! DB immediately.
-//!
-//! ## Encryption guard
-//!
-//! Keys are created on first use, but **only when the document store encrypts
-//! at rest** ([`crate::db::store::DocumentStore::is_encrypted`]). In the dev
-//! plaintext store no key is created and the resolver returns `None`, so
-//! callers fall back to the common key — a private key is never persisted in
-//! the clear, and there is no config flag threaded through call sites.
-
-use std::collections::HashMap;
-use std::sync::{Arc, Mutex};
-use std::time::{Duration, Instant};
+//! Staged rotate (a Next key must be published for [`PUBLISH_AHEAD_HOURS`]
+//! before promotion), gated revoke of Previous keys (token-drain window), and
+//! the emergency path that replaces the whole key set at once. All mutations
+//! are OCC transactions guarded by the org's claimed subdomain.
 
 use anyhow::{Context, Result};
-use axum::http::StatusCode;
-use base64::Engine;
-use base64::engine::general_purpose::STANDARD;
 use jiff::{Span, Timestamp};
-use secrecy::ExposeSecret;
-use zeroize::Zeroizing;
 
+use super::{KeyMaterial, ensure_key, generate_key_material};
 use crate::AppState;
-use crate::config::ServerConfig;
 use crate::db::audit::AuditStore;
-use crate::db::document_type::Document;
 use crate::db::documents::oauth::JwsAlgorithm;
 use crate::db::documents::organization::{OrgSigningKeyDoc, OrganizationDoc, SigningKeyState};
-use crate::db::store::{DocumentStore, StoreTransaction};
-use crate::db::{self, Organization};
-use crate::error::ServiceError;
-
-use super::discovery::{JwksResponse, build_jwks};
-use crate::crypto::keys::{Jwk, OidcRsaSigningKey, OidcSigningKey};
+use crate::db::store::StoreTransaction;
+use crate::db::{self};
 
 /// Hours a Next key must have been published in the JWKS before an operator
 /// rotate may promote it to the signer.
@@ -106,390 +39,6 @@ const RETIREMENT_FLOOR_HOURS: i64 = 8;
 /// (1h), plus a conservative buffer. The gate formula is:
 /// `demoted_at + max(session_hours, RETIREMENT_FLOOR_HOURS) + RETIREMENT_MARGIN_HOURS`.
 const RETIREMENT_MARGIN_HOURS: i64 = 2;
-
-/// The issuer an org's federation tokens must use: its claimed subdomain
-/// issuer, or the shared base URL when it has no subdomain.
-///
-/// **Fail-closed:** an org that claimed a subdomain must never fall back to the
-/// shared issuer — its tokens would not match its discovery document or the
-/// relying party's OIDC provider config. A subdomain that can't be turned into
-/// an issuer URL is an error, not a silent downgrade.
-///
-/// Issuer selection is gated on whether a subdomain is *claimed*, not on
-/// at-rest encryption — unlike [`resolve_org_keys`], which yields a per-org
-/// signing key only under encryption. These stay consistent because claiming
-/// requires an encrypted store (enforced at the claim handler and the startup
-/// guard) and `is_encrypted()` is fixed for the life of the process: a claimed
-/// subdomain therefore implies per-org keys exist, so the org issuer is always
-/// backed by an org-specific key. The one exception — org issuer served under
-/// the shared key — requires an out-of-band DB write on a plaintext deployment;
-/// [`resolve_org_keys`]'s encryption check is the fail-safe backstop for it.
-///
-/// # Errors
-/// Returns `ServiceError` if a claimed label can't be built into an issuer URL.
-pub fn org_issuer_or_base(
-    config: &ServerConfig,
-    org: Option<&Organization>,
-) -> Result<String, ServiceError> {
-    if let Some(label) = org.and_then(|o| o.subdomain.as_deref()) {
-        config.org_issuer(label).ok_or_else(|| {
-            tracing::error!("could not build org issuer for label '{label}' from base_url");
-            ServiceError::api(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "issuer_error",
-                "Failed to construct the organization issuer",
-            )
-        })
-    } else {
-        Ok(config.base_url.clone())
-    }
-}
-
-/// An org's active signing keys — one per algorithm.
-///
-/// Callers project the field they need (`&keys.es256` / `&keys.rs256`).
-pub struct OrgKeys {
-    pub es256: OidcSigningKey,
-    pub rs256: OidcRsaSigningKey,
-}
-
-/// A unified snapshot of an org's full key set, held by [`OrgKeysCache`].
-///
-/// Both projections (signing and JWKS) derive from the same `list_all("org_id")`
-/// result, so the signing key is always in this instance's own published JWKS —
-/// no cross-projection staleness within a single server instance.
-pub struct OrgKeySetSnapshot {
-    /// Current signers for token issuance — ES256 and RS256.
-    pub signers: OrgKeys,
-    /// Ordered public JWKs for all live keys (Current + Next + Previous).
-    ///
-    /// Sorted: all RS256 keys first (Current → Next → Previous),
-    /// then all ES256 keys in the same order. Per OIDC Core §3.1.3.7 RS256
-    /// must appear first; within an algorithm the order is deterministic for
-    /// stable response bytes and test assertions.
-    pub jwks: Vec<Jwk>,
-}
-
-/// How long a resolved key set may be served from [`OrgKeysCache`].
-const ORG_KEYS_CACHE_TTL: Duration = Duration::from_mins(1);
-
-/// Inner map type for `OrgKeysCache`: org ID → (insert time, snapshot).
-type OrgKeysCacheMap = HashMap<String, (Instant, Arc<OrgKeySetSnapshot>)>;
-
-/// Cache of resolved per-org key set snapshots, keyed by org ID.
-///
-/// The inner `Arc<Mutex<…>>` makes `OrgKeysCache` cheaply cloneable: the
-/// background cleanup task and request handlers share the same underlying map
-/// so cache invalidations propagate across both.
-///
-/// Every rotation transition must call `invalidate(org_id)` so the next
-/// request after a state change rebuilds from the DB rather than serving a
-/// stale snapshot.
-#[derive(Default)]
-pub struct OrgKeysCache {
-    entries: Arc<Mutex<OrgKeysCacheMap>>,
-}
-
-impl Clone for OrgKeysCache {
-    fn clone(&self) -> Self {
-        Self {
-            entries: Arc::clone(&self.entries),
-        }
-    }
-}
-
-impl OrgKeysCache {
-    /// Return a cached snapshot for `org_id`, if present and not expired.
-    fn get(&self, org_id: &str) -> Option<Arc<OrgKeySetSnapshot>> {
-        let Ok(map) = self.entries.lock() else {
-            return None;
-        };
-        map.get(org_id)
-            .filter(|(inserted_at, _)| inserted_at.elapsed() < ORG_KEYS_CACHE_TTL)
-            .map(|(_, snap)| Arc::clone(snap))
-    }
-
-    /// Cache `snapshot`, pruning expired entries so deleted orgs don't accumulate.
-    fn insert(&self, org_id: &str, snapshot: Arc<OrgKeySetSnapshot>) {
-        let Ok(mut map) = self.entries.lock() else {
-            return;
-        };
-        map.retain(|_, (inserted_at, _)| inserted_at.elapsed() < ORG_KEYS_CACHE_TTL);
-        map.insert(org_id.to_string(), (Instant::now(), snapshot));
-    }
-
-    /// Evict the cached snapshot for `org_id`.
-    ///
-    /// Called after every rotation transition (rotate, revoke, emergency) so
-    /// the next request rebuilds from the DB immediately rather than serving
-    /// a now-stale snapshot.
-    pub fn invalidate(&self, org_id: &str) {
-        if let Ok(mut map) = self.entries.lock() {
-            map.remove(org_id);
-        }
-    }
-}
-
-/// Resolve an org's own signing key set, creating it on first use.
-///
-/// Returns `None` when the org has no claimed subdomain, or when the document
-/// store doesn't encrypt at rest — the caller then falls back to the common
-/// platform key. Resolutions are served from a per-org cache for
-/// [`ORG_KEYS_CACHE_TTL`], so the token hot paths don't re-read and unseal key
-/// rows on every request.
-///
-/// The returned snapshot also contains the ordered public-JWK list for all live
-/// keys (Current + Next + Previous) which [`org_jwks`] uses directly.
-///
-/// # Errors
-/// Returns an error if key creation or loading fails.
-pub async fn resolve_org_keys(
-    state: &Arc<AppState>,
-    org: Option<&Organization>,
-) -> Result<Option<Arc<OrgKeySetSnapshot>>> {
-    let Some(org) = org else { return Ok(None) };
-    if org.subdomain.is_none() {
-        // Self-heal: release cancels rotation keys in the DB, but the release
-        // paths live in the db layer and cannot reach this cache. Purging on
-        // the first resolve for a released org keeps a quick reclaim from
-        // resurrecting a pre-release snapshot.
-        state.org_keys_cache.invalidate(&org.id);
-        return Ok(None);
-    }
-    if !state.store.is_encrypted() {
-        return Ok(None);
-    }
-    if let Some(snap) = state.org_keys_cache.get(&org.id) {
-        return Ok(Some(snap));
-    }
-    let store = &state.store;
-
-    // Single list_all call to check which keys already exist (avoids
-    // serial round-trips by discovering the full doc set upfront). The Auth0
-    // invariant is that a claimed org always has a Current signer AND a
-    // pre-staged Next successor per algorithm, so both are created here.
-    let docs = db::list_org_signing_keys(store, &org.id).await?;
-    let has = |alg: JwsAlgorithm, state: SigningKeyState| {
-        docs.iter()
-            .any(|d| d.data.alg == alg && d.data.state == state)
-    };
-    let mut created = false;
-    for alg in [JwsAlgorithm::Es256, JwsAlgorithm::Rs256] {
-        for state in [SigningKeyState::Current, SigningKeyState::Next] {
-            if !has(alg, state) {
-                ensure_key(store, &org.id, alg, state).await?;
-                created = true;
-            }
-        }
-    }
-
-    // Re-read only when we just generated new keys; otherwise build from the
-    // already-loaded list (saves the extra round-trip in the common case).
-    let docs = if created {
-        db::list_org_signing_keys(store, &org.id).await?
-    } else {
-        docs
-    };
-
-    let Some(snap) = build_snapshot(&docs)? else {
-        return Ok(None);
-    };
-    let snap = Arc::new(snap);
-    state.org_keys_cache.insert(&org.id, Arc::clone(&snap));
-    Ok(Some(snap))
-}
-
-/// Build the JWKS served on `org`'s issuer-subdomain host: the org's own keys,
-/// or the common keys when the org has none (dev / not-yet-encrypted). RSA
-/// first (OIDC Core §3.1.3.7), then EC; within each alg: Current → Next →
-/// Previous.
-///
-/// Uses the unified cache snapshot so signing and JWKS are always consistent
-/// within an instance.
-///
-/// # Errors
-/// Returns `ServiceError` if a key cannot be resolved or exported.
-pub async fn org_jwks(
-    state: &Arc<AppState>,
-    org: &Organization,
-) -> Result<JwksResponse, ServiceError> {
-    let Some(snap) = resolve_org_keys(state, Some(org))
-        .await
-        .map_err(|e| ServiceError::Internal(format!("resolve org keys: {e}")))?
-    else {
-        return build_jwks(state);
-    };
-    Ok(JwksResponse {
-        keys: snap.jwks.clone(),
-    })
-}
-
-/// Freshly generated key material for one algorithm: PKCS#8 DER + derived kid.
-struct KeyMaterial {
-    der: Zeroizing<Vec<u8>>,
-    kid: String,
-}
-
-impl KeyMaterial {
-    /// Build an [`OrgSigningKeyDoc`] carrying this material in `state`.
-    ///
-    /// Borrows so the same material can be reused across OCC retries.
-    fn doc(&self, org_id: &str, alg: JwsAlgorithm, state: SigningKeyState) -> OrgSigningKeyDoc {
-        OrgSigningKeyDoc {
-            org_id: org_id.to_string(),
-            alg,
-            kid: self.kid.clone(),
-            private_pkcs8_der_b64: STANDARD.encode(&self.der).into(),
-            state,
-            staged_at: None,
-            demoted_at: None,
-        }
-    }
-}
-
-/// Generate fresh key material for `alg`.
-///
-/// RSA generation is expensive, so it runs on the blocking pool and callers
-/// keep it out of retry loops.
-async fn generate_key_material(alg: JwsAlgorithm) -> Result<KeyMaterial> {
-    match alg {
-        JwsAlgorithm::Es256 => {
-            let der = OidcSigningKey::generate_pkcs8_der().context("ES256 keygen")?;
-            let kid = OidcSigningKey::from_pkcs8_der(&der)
-                .context("ES256: parse DER")?
-                .key_id()
-                .to_string();
-            Ok(KeyMaterial { der, kid })
-        }
-        JwsAlgorithm::Rs256 => {
-            let der = tokio::task::spawn_blocking(OidcRsaSigningKey::generate_pkcs8_der)
-                .await
-                .context("RS256 keygen task panicked")??;
-            let kid = OidcRsaSigningKey::from_pkcs8_der(&der)
-                .context("RS256: parse DER")?
-                .key_id()
-                .to_string();
-            Ok(KeyMaterial { der, kid })
-        }
-        other => anyhow::bail!("generate_key_material: unsupported algorithm {other:?}"),
-    }
-}
-
-/// Create the key in `state` for `(org_id, alg)` if it doesn't exist.
-///
-/// `Next` keys record `staged_at = now` — the start of their publish window
-/// (the always-staged invariant: the successor is published in the JWKS from
-/// first use so relying-party caches are warm long before any rotate).
-/// Idempotent: the deterministic ID makes a concurrent creation or retry
-/// collide on the primary key rather than insert a duplicate.
-async fn ensure_key(
-    store: &DocumentStore,
-    org_id: &str,
-    alg: JwsAlgorithm,
-    state: SigningKeyState,
-) -> Result<()> {
-    if state == SigningKeyState::Previous {
-        anyhow::bail!("ensure_key: Previous keys are only created by rotation");
-    }
-    if db::get_org_signing_key(store, org_id, alg, state)
-        .await?
-        .is_some()
-    {
-        return Ok(());
-    }
-    let doc = OrgSigningKeyDoc {
-        staged_at: (state == SigningKeyState::Next).then(Timestamp::now),
-        ..generate_key_material(alg).await?.doc(org_id, alg, state)
-    };
-    db::try_insert_org_signing_key(store, &doc).await?;
-    Ok(())
-}
-
-/// Decode the DER from an `OrgSigningKeyDoc` into a `Zeroizing<Vec<u8>>`.
-fn decode_der(doc: &OrgSigningKeyDoc) -> Result<Zeroizing<Vec<u8>>> {
-    Ok(Zeroizing::new(
-        STANDARD
-            .decode(doc.private_pkcs8_der_b64.expose_secret())
-            .context("decode org signing key DER")?,
-    ))
-}
-
-/// State-priority ordering used for JWKS sort within an algorithm.
-fn state_priority(state: SigningKeyState) -> u8 {
-    match state {
-        SigningKeyState::Current => 0,
-        SigningKeyState::Next => 1,
-        SigningKeyState::Previous => 2,
-    }
-}
-
-/// Build a unified [`OrgKeySetSnapshot`] from a pre-loaded slice of key docs.
-///
-/// Finds the Current ES256 and RS256 keys for signing, and builds the sorted
-/// public-JWK list for the JWKS endpoint. Returns `None` if no Current key is
-/// found for either algorithm.
-///
-/// **Ordering:** RS256 keys first, then ES256; within each algorithm:
-/// Current → Next → Previous. This is a hard constraint for OIDC Core
-/// §3.1.3.7 compliance (RS256 must appear first).
-///
-/// Takes pre-loaded docs from the caller (avoids an extra DB round-trip
-/// when `resolve_org_keys` already has the full doc list in hand).
-fn build_snapshot(docs: &[Document<OrgSigningKeyDoc>]) -> Result<Option<OrgKeySetSnapshot>> {
-    // Separate by alg and sort by state priority within each alg.
-    let mut rs256_docs: Vec<_> = docs
-        .iter()
-        .filter(|d| d.data.alg == JwsAlgorithm::Rs256)
-        .collect();
-    let mut es256_docs: Vec<_> = docs
-        .iter()
-        .filter(|d| d.data.alg == JwsAlgorithm::Es256)
-        .collect();
-    rs256_docs.sort_by_key(|d| state_priority(d.data.state));
-    es256_docs.sort_by_key(|d| state_priority(d.data.state));
-
-    // Find the Current key for each alg — none means the org has no live
-    // signing capability; return None to fall back to the common key.
-    let Some(current_rs256_doc) = rs256_docs
-        .iter()
-        .find(|d| d.data.state == SigningKeyState::Current)
-    else {
-        return Ok(None);
-    };
-    let Some(current_es256_doc) = es256_docs
-        .iter()
-        .find(|d| d.data.state == SigningKeyState::Current)
-    else {
-        return Ok(None);
-    };
-
-    let rs256_der = decode_der(&current_rs256_doc.data)?;
-    let es256_der = decode_der(&current_es256_doc.data)?;
-    let rs256 = OidcRsaSigningKey::from_pkcs8_der(&rs256_der)?;
-    let es256 = OidcSigningKey::from_pkcs8_der(&es256_der)?;
-
-    // Build public JWKs: RS256 first (all states), then ES256 (all states).
-    let mut jwks = Vec::with_capacity(rs256_docs.len().saturating_add(es256_docs.len()));
-    for doc in &rs256_docs {
-        let der = decode_der(&doc.data)?;
-        let key = OidcRsaSigningKey::from_pkcs8_der(&der)?;
-        jwks.push(Jwk::Rsa(key.public_key_jwk()?));
-    }
-    for doc in &es256_docs {
-        let der = decode_der(&doc.data)?;
-        let key = OidcSigningKey::from_pkcs8_der(&der)?;
-        jwks.push(Jwk::Ec(key.public_key_jwk()?));
-    }
-
-    Ok(Some(OrgKeySetSnapshot {
-        signers: OrgKeys { es256, rs256 },
-        jwks,
-    }))
-}
-
-// ---------------------------------------------------------------------------
-// Rotation error type (follows the OrgCasError pattern)
-// ---------------------------------------------------------------------------
 
 /// Error type for rotation operations that may be OCC-retried.
 ///
@@ -611,7 +160,7 @@ async fn audit_best_effort(
 
 /// The instant a key staged at `staged_at` has been published long enough for
 /// relying-party JWKS caches to have seen its kid.
-fn publish_ready_at(staged_at: Timestamp) -> Result<Timestamp> {
+pub(super) fn publish_ready_at(staged_at: Timestamp) -> Result<Timestamp> {
     staged_at
         .checked_add(Span::new().hours(PUBLISH_AHEAD_HOURS))
         .context("publish window overflow")
@@ -626,7 +175,7 @@ const REVOKE_GATE_CAP_HOURS: i64 = 24 * 365;
 /// gate `max(session_hours, RETIREMENT_FLOOR_HOURS) + RETIREMENT_MARGIN_HOURS`
 /// (the floor keeps a `session_hours` reduction from shortening the gate under
 /// what already-issued tokens need), capped at [`REVOKE_GATE_CAP_HOURS`].
-fn revoke_ready_at(demoted_at: Timestamp, session_hours: u64) -> Result<Timestamp> {
+pub(super) fn revoke_ready_at(demoted_at: Timestamp, session_hours: u64) -> Result<Timestamp> {
     // An unrepresentable session_hours clamps LONG (to the cap), never short:
     // deleting a key early breaks live sessions; waiting longer breaks nothing.
     let session_h = i64::try_from(session_hours).unwrap_or(REVOKE_GATE_CAP_HOURS);
@@ -1012,130 +561,6 @@ pub async fn revoke_org_previous_keys(
 }
 
 // ---------------------------------------------------------------------------
-// Admin page panel
-// ---------------------------------------------------------------------------
-
-/// One key row for the admin page's signing-key panel.
-pub(crate) struct OrgKeyPanelRow {
-    pub alg: JwsAlgorithm,
-    pub state: SigningKeyState,
-    pub kid: String,
-    /// `staged_at` for Next keys, `demoted_at` for Previous keys.
-    pub since: Option<Timestamp>,
-}
-
-/// The admin page's view of an org's key set and what the two buttons may do.
-///
-/// Blocked reasons reuse the operation outcome types so the page and the POST
-/// handlers derive their messages from the same source.
-pub(crate) struct OrgKeyPanel {
-    /// RS256 rows first, Current → Next → Previous within each algorithm.
-    pub rows: Vec<OrgKeyPanelRow>,
-    /// Why rotate would be rejected right now; `None` when it may proceed.
-    pub rotate_blocked: Option<RotateOutcome>,
-    /// Why revoke would be rejected right now; `None` when it may proceed.
-    pub revoke_blocked: Option<RevokeOutcome>,
-}
-
-/// Build the signing-key panel for the admin subdomain page.
-///
-/// Read-only: unlike the operations themselves this never stages or heals
-/// anything, so rendering the page has no side effects.
-///
-/// # Errors
-/// Returns an error if the key list cannot be loaded or a stored key is
-/// missing its state timestamp.
-pub(crate) async fn org_key_panel(state: &AppState, org_id: &str) -> Result<OrgKeyPanel> {
-    let now = Timestamp::now();
-    let mut docs = db::list_org_signing_keys(&state.store, org_id).await?;
-    docs.sort_by_key(|d| {
-        (
-            d.data.alg != JwsAlgorithm::Rs256,
-            state_priority(d.data.state),
-        )
-    });
-
-    let mut has_previous = false;
-    let mut missing_current_or_next = false;
-    let mut latest_next_ready: Option<Timestamp> = None;
-    let mut latest_revoke_ready: Option<Timestamp> = None;
-
-    for alg in [JwsAlgorithm::Es256, JwsAlgorithm::Rs256] {
-        let find = |state: SigningKeyState| {
-            docs.iter()
-                .find(|d| d.data.alg == alg && d.data.state == state)
-        };
-        if let Some(prev) = find(SigningKeyState::Previous) {
-            has_previous = true;
-            let demoted_at = prev
-                .data
-                .demoted_at
-                .context("previous key is missing its demoted_at timestamp")?;
-            let ready_at = revoke_ready_at(demoted_at, state.config().session_hours)?;
-            if latest_revoke_ready.is_none_or(|cur| ready_at > cur) {
-                latest_revoke_ready = Some(ready_at);
-            }
-        }
-        match (find(SigningKeyState::Current), find(SigningKeyState::Next)) {
-            (Some(_), Some(next)) => {
-                let staged_at = next
-                    .data
-                    .staged_at
-                    .context("next key is missing its staged_at timestamp")?;
-                let ready_at = publish_ready_at(staged_at)?;
-                if latest_next_ready.is_none_or(|cur| ready_at > cur) {
-                    latest_next_ready = Some(ready_at);
-                }
-            }
-            // Legacy rows: a rotate attempt heals the missing Next and starts
-            // its publish window, so report the state a rotate would produce.
-            (Some(_), None) => {
-                let ready_at = publish_ready_at(now)?;
-                if latest_next_ready.is_none_or(|cur| ready_at > cur) {
-                    latest_next_ready = Some(ready_at);
-                }
-            }
-            // Keys are created on first use; until then nothing can rotate.
-            (None, _) => missing_current_or_next = true,
-        }
-    }
-
-    let rotate_blocked = if has_previous {
-        Some(RotateOutcome::PreviousUnrevoked)
-    } else if missing_current_or_next {
-        Some(RotateOutcome::NotBootstrapped)
-    } else {
-        match latest_next_ready {
-            Some(ready_at) if ready_at > now => Some(RotateOutcome::NextNotReady { ready_at }),
-            _ => None,
-        }
-    };
-    let revoke_blocked = if has_previous {
-        match latest_revoke_ready {
-            Some(ready_at) if ready_at > now => Some(RevokeOutcome::NotReady { ready_at }),
-            _ => None,
-        }
-    } else {
-        Some(RevokeOutcome::NothingToRevoke)
-    };
-
-    let rows = docs
-        .into_iter()
-        .map(|d| OrgKeyPanelRow {
-            alg: d.data.alg,
-            state: d.data.state,
-            kid: d.data.kid,
-            since: d.data.staged_at.or(d.data.demoted_at),
-        })
-        .collect();
-    Ok(OrgKeyPanel {
-        rows,
-        rotate_blocked,
-        revoke_blocked,
-    })
-}
-
-// ---------------------------------------------------------------------------
 // Emergency rotation — replace the whole key set at once
 // ---------------------------------------------------------------------------
 
@@ -1288,84 +713,20 @@ pub async fn emergency_rotate_org_keys(
 #[cfg(test)]
 #[expect(
     clippy::unwrap_used,
-    clippy::expect_used,
     clippy::panic,
     reason = "test code: panic on assertion failure is acceptable"
 )]
 mod tests {
-    use std::sync::Arc;
-
     use jiff::{Span, Timestamp};
 
-    use super::{
-        EmergencyOutcome, Operator, PUBLISH_AHEAD_HOURS, RETIREMENT_FLOOR_HOURS,
-        RETIREMENT_MARGIN_HOURS, RevokeOutcome, RotateOutcome, emergency_rotate_org_keys,
-        org_issuer_or_base, revoke_org_previous_keys, rotate_org_keys, state_priority,
-    };
-    use crate::crypto::keys::Jwk;
+    use super::super::test_support::{NO_OPERATOR, backdate, setup};
+    use super::*;
     use crate::db::documents::organization::SigningKeyState;
     use crate::db::{
-        JwsAlgorithm, OrgSigningKeyDoc, Organization, claim_subdomain, deterministic_org_key_id,
-        get_org_signing_key, release_subdomain,
+        JwsAlgorithm, OrgSigningKeyDoc, deterministic_org_key_id, get_org_signing_key,
+        release_subdomain,
     };
     use crate::services::oidc::resolve_org_keys;
-    use crate::test_utils::{create_test_org, test_app_state_encrypted, test_config};
-
-    /// No operator identity — service-level tests don't exercise audit trails.
-    const NO_OPERATOR: Operator<'static> = Operator {
-        user_id: None,
-        email: None,
-    };
-
-    fn org(subdomain: Option<&str>) -> Organization {
-        Organization {
-            id: "org1".to_string(),
-            domain: "acme.com".to_string(),
-            name: None,
-            created_at: Timestamp::now(),
-            created_by_user_id: None,
-            additional_domains: Vec::new(),
-            subdomain: subdomain.map(String::from),
-        }
-    }
-
-    #[test]
-    fn issuer_is_org_subdomain_or_base_url() {
-        let config = test_config();
-        // No org, or an org without a subdomain -> shared base URL.
-        assert_eq!(org_issuer_or_base(&config, None).unwrap(), config.base_url);
-        assert_eq!(
-            org_issuer_or_base(&config, Some(&org(None))).unwrap(),
-            config.base_url
-        );
-        // A claimed subdomain -> the org issuer host.
-        assert_eq!(
-            org_issuer_or_base(&config, Some(&org(Some("acme")))).unwrap(),
-            "https://acme.test.example.com"
-        );
-    }
-
-    #[test]
-    fn state_priority_orders_current_next_previous() {
-        assert!(state_priority(SigningKeyState::Current) < state_priority(SigningKeyState::Next));
-        assert!(state_priority(SigningKeyState::Next) < state_priority(SigningKeyState::Previous));
-    }
-
-    #[test]
-    fn org_key_state_serde_round_trip_and_legacy_default() {
-        for state in [
-            SigningKeyState::Current,
-            SigningKeyState::Next,
-            SigningKeyState::Previous,
-        ] {
-            let json = serde_json::to_string(&state).expect("serialise");
-            let back: SigningKeyState = serde_json::from_str(&json).expect("deserialise");
-            assert_eq!(state, back, "round-trip failed for {json}");
-        }
-        // Rows written before rotation existed have no `state` field and must
-        // deserialize as Current.
-        assert_eq!(SigningKeyState::default(), SigningKeyState::Current);
-    }
 
     #[test]
     fn publish_window_is_the_locked_twenty_four_hours() {
@@ -1445,155 +806,6 @@ mod tests {
             )
             .unwrap();
         assert_eq!(super::revoke_ready_at(demoted_at, 0).unwrap(), floor);
-    }
-
-    /// The Current ID must be byte-identical to the deterministic ID used
-    /// before rotation existed so old rows keep their IDs without migration.
-    #[test]
-    fn current_id_matches_pre_rotation_formula() {
-        use aws_lc_rs::digest::{self, SHA256};
-
-        let org_id = "migration-parity-test-org";
-        let alg = JwsAlgorithm::Es256;
-
-        let mut ctx = digest::Context::new(&SHA256);
-        ctx.update(b"org_signing_key\0");
-        ctx.update(org_id.as_bytes());
-        ctx.update(b"\0");
-        ctx.update(alg.as_str().as_bytes());
-        let expected = hex::encode(ctx.finish().as_ref());
-
-        assert_eq!(
-            deterministic_org_key_id(org_id, alg, SigningKeyState::Current),
-            expected
-        );
-    }
-
-    #[test]
-    fn key_ids_are_distinct_per_state_alg_and_org() {
-        for alg in [JwsAlgorithm::Es256, JwsAlgorithm::Rs256] {
-            let ids = [
-                deterministic_org_key_id("test-org", alg, SigningKeyState::Current),
-                deterministic_org_key_id("test-org", alg, SigningKeyState::Next),
-                deterministic_org_key_id("test-org", alg, SigningKeyState::Previous),
-            ];
-            let unique: std::collections::HashSet<&String> = ids.iter().collect();
-            assert_eq!(unique.len(), 3, "{alg:?}: state IDs collide: {ids:?}");
-            // Deterministic: the same inputs produce the same ID.
-            assert_eq!(
-                deterministic_org_key_id("test-org", alg, SigningKeyState::Current),
-                ids[0]
-            );
-        }
-        assert_ne!(
-            deterministic_org_key_id("org-a", JwsAlgorithm::Es256, SigningKeyState::Current),
-            deterministic_org_key_id("org-b", JwsAlgorithm::Es256, SigningKeyState::Current),
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // Async tests against an encrypted store
-    // -----------------------------------------------------------------------
-
-    /// Encrypted AppState with a fresh org holding a claimed subdomain.
-    async fn setup() -> (Arc<crate::AppState>, String, Organization) {
-        let state = test_app_state_encrypted().await;
-        let org = create_test_org(&state.store, "acme.com").await;
-        claim_subdomain(&state.store, &org.id, "acme-com")
-            .await
-            .unwrap();
-        let org = crate::db::get_organization(&state.store, &org.id)
-            .await
-            .unwrap()
-            .unwrap();
-        let org_id = org.id.clone();
-        (state, org_id, org)
-    }
-
-    /// Rewrite a key's publish/demotion timestamp so gate tests don't sleep.
-    async fn backdate(
-        state: &crate::AppState,
-        org_id: &str,
-        alg: JwsAlgorithm,
-        key_state: SigningKeyState,
-        hours: i64,
-    ) {
-        let id = deterministic_org_key_id(org_id, alg, key_state);
-        let doc = state
-            .store
-            .get::<OrgSigningKeyDoc>(&id)
-            .await
-            .unwrap()
-            .expect("key to backdate must exist");
-        let past = Timestamp::now()
-            .checked_sub(Span::new().hours(hours))
-            .unwrap();
-        let mut data = doc.data;
-        match key_state {
-            SigningKeyState::Next => data.staged_at = Some(past),
-            SigningKeyState::Previous => data.demoted_at = Some(past),
-            SigningKeyState::Current => panic!("current keys have no timestamp"),
-        }
-        state.store.update(&id, &data).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn first_use_creates_current_and_next_and_signs_with_current() {
-        let (state, org_id, org) = setup().await;
-
-        let snap = resolve_org_keys(&state, Some(&org)).await.unwrap().unwrap();
-        // Two algorithms x (Current + Next) published from day one.
-        assert_eq!(snap.jwks.len(), 4, "expected Current+Next for both algs");
-
-        // The signer is the Current key, not the staged Next.
-        let current = get_org_signing_key(
-            &state.store,
-            &org_id,
-            JwsAlgorithm::Es256,
-            SigningKeyState::Current,
-        )
-        .await
-        .unwrap()
-        .unwrap();
-        assert_eq!(snap.signers.es256.key_id(), current.data.kid);
-
-        let next = get_org_signing_key(
-            &state.store,
-            &org_id,
-            JwsAlgorithm::Es256,
-            SigningKeyState::Next,
-        )
-        .await
-        .unwrap()
-        .unwrap();
-        assert!(next.data.staged_at.is_some(), "next key records staged_at");
-        assert_ne!(next.data.kid, current.data.kid);
-    }
-
-    #[tokio::test]
-    async fn jwks_lists_rsa_first_with_distinct_kids() {
-        let (state, _org_id, org) = setup().await;
-        let snap = resolve_org_keys(&state, Some(&org)).await.unwrap().unwrap();
-
-        let mut saw_ec = false;
-        for jwk in &snap.jwks {
-            match jwk {
-                Jwk::Rsa(_) => assert!(!saw_ec, "RSA JWK after an EC JWK"),
-                Jwk::Ec(_) => saw_ec = true,
-            }
-        }
-        assert!(saw_ec, "at least one EC JWK must be present");
-
-        let kids: Vec<&str> = snap
-            .jwks
-            .iter()
-            .map(|jwk| match jwk {
-                Jwk::Rsa(rsa) => rsa.kid.as_str(),
-                Jwk::Ec(ec) => ec.kid.as_str(),
-            })
-            .collect();
-        let unique: std::collections::HashSet<&str> = kids.iter().copied().collect();
-        assert_eq!(kids.len(), unique.len(), "duplicate kids: {kids:?}");
     }
 
     #[tokio::test]
@@ -1831,83 +1043,6 @@ mod tests {
         let result = revoke_org_previous_keys(&state, &org_id, NO_OPERATOR).await;
         let msg = result.unwrap_err().to_string();
         assert!(msg.contains("demoted_at"), "got: {msg}");
-    }
-
-    #[tokio::test]
-    async fn panel_gates_follow_the_key_set_state() {
-        let (state, org_id, org) = setup().await;
-
-        // Before first use: nothing to rotate or revoke.
-        let panel = super::org_key_panel(&state, &org_id).await.unwrap();
-        assert!(panel.rows.is_empty());
-        assert_eq!(panel.rotate_blocked, Some(RotateOutcome::NotBootstrapped));
-        assert_eq!(panel.revoke_blocked, Some(RevokeOutcome::NothingToRevoke));
-
-        // Bootstrapped, next keys fresh: rotate gated by the publish window.
-        resolve_org_keys(&state, Some(&org)).await.unwrap();
-        let panel = super::org_key_panel(&state, &org_id).await.unwrap();
-        assert_eq!(panel.rows.len(), 4);
-        assert!(
-            matches!(
-                panel.rotate_blocked,
-                Some(RotateOutcome::NextNotReady { .. })
-            ),
-            "got {:?}",
-            panel.rotate_blocked
-        );
-
-        // Aged next keys: rotate allowed.
-        backdate(
-            &state,
-            &org_id,
-            JwsAlgorithm::Es256,
-            SigningKeyState::Next,
-            25,
-        )
-        .await;
-        backdate(
-            &state,
-            &org_id,
-            JwsAlgorithm::Rs256,
-            SigningKeyState::Next,
-            25,
-        )
-        .await;
-        let panel = super::org_key_panel(&state, &org_id).await.unwrap();
-        assert_eq!(panel.rotate_blocked, None);
-
-        // After a rotate: the unrevoked previous outranks the (young) next
-        // key's warm-up as the rotate-blocked reason, and revoke is gated by
-        // the drain window.
-        rotate_org_keys(&state, &org_id, NO_OPERATOR).await.unwrap();
-        let panel = super::org_key_panel(&state, &org_id).await.unwrap();
-        assert_eq!(panel.rows.len(), 6);
-        assert_eq!(panel.rotate_blocked, Some(RotateOutcome::PreviousUnrevoked));
-        assert!(
-            matches!(panel.revoke_blocked, Some(RevokeOutcome::NotReady { .. })),
-            "got {:?}",
-            panel.revoke_blocked
-        );
-
-        // Past the drain window: revoke allowed.
-        backdate(
-            &state,
-            &org_id,
-            JwsAlgorithm::Es256,
-            SigningKeyState::Previous,
-            30,
-        )
-        .await;
-        backdate(
-            &state,
-            &org_id,
-            JwsAlgorithm::Rs256,
-            SigningKeyState::Previous,
-            30,
-        )
-        .await;
-        let panel = super::org_key_panel(&state, &org_id).await.unwrap();
-        assert_eq!(panel.revoke_blocked, None);
     }
 
     #[tokio::test]
